Add configuration dialog
This is probably very suboptimal. I feel like the wrong things are
"holding" other things.

A connection should probably not need to be passed around to everything.

A refcel to the Connection is given to the ConversationList and each
Conversation. It uses the Connection to do two things:

1. Know the self-address
2. Send commands

Things that need to know self-address should probably observe the
configuration for changes to it.

There should be some way to pass around the "command sender," which can
survive reconfig/reconnects.

If those two things are done, the Connection can probably be the god
object in the hierarchy. It can be the object that handles events, and
then delegates them down the tree.