Reduce mutability
Because of how controllers control views, observables don't need to notify mutable observers.

That is, the controllers don't actually contain any state that needs to be mutated when they
receive a notification from a model they observe. They just have to mutate their views, and thanks
to GTK's weird pattern of doing everything with borrowed refs, they don't have to be mut.

If controllers start hanging onto state, this would change, but I'll try to avoid that.